Move edit file code into its own class
This was really starting to clog up Main.cs.
